Repackage Clerk-related DTOs to dedicated module
