+++ api/src/main/java/com/example/echo_api/modules/clerk/dto/sdk/ClerkUserDTO.java
@@ -1,4 +1,4 @@
-package com.example.echo_api.persistence.dto.adapter;
+package com.example.echo_api.modules.clerk.dto.sdk;
 
 import java.util.Map;
