+++ api/src/main/java/com/example/echo_api/modules/clerk/dto/webhook/ClerkWebhookEvent.java
@@ -1,4 +1,4 @@
-package com.example.echo_api.persistence.dto.request.clerk.webhook;
+package com.example.echo_api.modules.clerk.dto.webhook;
 
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
