+++ api/src/main/java/com/example/echo_api/modules/clerk/dto/webhook/ClerkWebhookEventData.java
@@ -1,0 +1,3 @@
+package com.example.echo_api.modules.clerk.dto.webhook;
+
+public interface ClerkWebhookEventData {}
