+++ api/src/main/java/com/example/echo_api/modules/clerk/service/ClerkSyncService.java
@@ -1,7 +1,7 @@
 package com.example.echo_api.modules.clerk.service;
 
 import com.example.echo_api.exception.custom.internalserver.ClerkException;
-import com.example.echo_api.persistence.dto.request.clerk.webhook.ClerkWebhookEvent;
+import com.example.echo_api.modules.clerk.dto.webhook.ClerkWebhookEvent;
 import com.example.echo_api.persistence.model.user.User;
 
 public interface ClerkSyncService {
