+++ api/src/main/java/com/example/echo_api/modules/clerk/service/ClerkSyncServiceImpl.java
@@ -4,10 +4,10 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.clerk.backend_api.Clerk;
-import com.example.echo_api.persistence.dto.adapter.ClerkUserDTO;
-import com.example.echo_api.persistence.dto.request.clerk.webhook.ClerkWebhookEvent;
-import com.example.echo_api.persistence.dto.request.clerk.webhook.data.UserDelete;
-import com.example.echo_api.persistence.dto.request.clerk.webhook.data.UserUpsert;
+import com.example.echo_api.modules.clerk.dto.sdk.ClerkUserDTO;
+import com.example.echo_api.modules.clerk.dto.webhook.ClerkWebhookEvent;
+import com.example.echo_api.modules.clerk.dto.webhook.UserDelete;
+import com.example.echo_api.modules.clerk.dto.webhook.UserUpsert;
 import com.example.echo_api.persistence.model.user.User;
 import com.example.echo_api.service.user.UserService;
 import com.example.echo_api.shared.service.SessionService;
