--- api/src/main/java/com/example/echo_api/persistence/dto/request/clerk/webhook/ClerkWebhookEventData.java
@@ -1,3 +1,0 @@
-package com.example.echo_api.persistence.dto.request.clerk.webhook;
-
-public interface ClerkWebhookEventData {}
