+++ api/src/main/java/com/example/echo_api/service/dev/DevServiceImpl.java
@@ -6,8 +6,8 @@ import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import com.example.echo_api.modules.clerk.dto.sdk.ClerkUserDTO;
 import com.example.echo_api.modules.clerk.service.ClerkSdkService;
-import com.example.echo_api.persistence.dto.adapter.ClerkUserDTO;
 import com.example.echo_api.persistence.model.user.User;
 import com.example.echo_api.service.user.UserService;
