+++ api/src/test/java/com/example/echo_api/integration/util/ClerkTestUtils.java
@@ -15,8 +15,8 @@ import com.clerk.backend_api.models.operations.CreateUserResponse;
 import com.clerk.backend_api.models.operations.UpdateUserRequestBody;
 import com.example.echo_api.config.ClerkConfig;
 import com.example.echo_api.exception.custom.internalserver.ClerkException;
-import com.example.echo_api.persistence.dto.adapter.ClerkUserDTO;
-import com.example.echo_api.persistence.mapper.ClerkUserMapper;
+import com.example.echo_api.modules.clerk.dto.sdk.ClerkUserDTO;
+import com.example.echo_api.modules.clerk.mapper.ClerkUserMapper;
 
 @Component
 public class ClerkTestUtils {
