+++ api/src/test/java/com/example/echo_api/modules/clerk/service/ClerkWebhookServiceTest.java
@@ -14,7 +14,7 @@ import org.springframework.http.HttpHeaders;
 
 import com.example.echo_api.exception.custom.badrequest.DeserializationException;
 import com.example.echo_api.exception.custom.unauthorised.WebhookVerificationException;
-import com.example.echo_api.persistence.dto.request.clerk.webhook.ClerkWebhookEvent;
+import com.example.echo_api.modules.clerk.dto.webhook.ClerkWebhookEvent;
 import com.example.echo_api.util.Utils;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.svix.Webhook;
